Separate Contact component from page route component
